requirements/base.txt
# base requirements

celery==4.4.7
chargebee==2.21.0
Django==3.2.14
django-braces==1.15.0
django-celery-beat==2.2.0
django-taggit==3.0.0
django-rq==2.5.1
psycopg2-binary==2.9.3
django-oauth-toolkit==2.1.0
mistune==2.0.4
djangorestframework==3.13.1
redis==4.3.4
Pillow==9.2.0
django-nested-admin==3.4.0
django-cors-headers==3.13.0
whitenoise==6.2.0
django-allauth==0.51.0
faker==13.15.0
django-filter==22.1
jsonmodels==2.5.1
djangorestframework-simplejwt==5.2.0
sentry-sdk==1.7.2
django-s3-storage==0.13.7
django-crum==0.7.9
django-redis==5.2.0

numpy==1.22.4
rq==1.13
djangorestframework-jwt==1.11.0
django-sesame==2.4
matplotlib==3.7.1
opencv-python==4.7.0.72
pycocotools==2.0.6
python-crontab==2.7.1
python-decouple==3.8
pytorch-lightning==1.9.0
torchmetrics==0.11.4
moviepy==1.0.3
filetype==1.2.0
wget==3.2
paddlepaddle
paddleocr>=2.0.1
stripe==5.4.0
langchain 
openai 
chromadb==0.3.29
PyPDF2
tiktoken 
